# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Feira das Ferramentas API",
    "version": "1.0.0",
    "description": "API REST completa para e-commerce de ferramentas com integração Mercado Pago e Melhor Envio.\n\n## Autenticação\nA maioria dos endpoints utiliza autenticação via cookie de sessão NextAuth. Endpoints de integração aceitam `X-INTERNAL-API-KEY` no header.",
    "contact": {
      "name": "Suporte API",
      "email": "suporte@feiradeferramentas.com.br"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento"
    },
    {
      "url": "https://feiradeferramentas.com.br",
      "description": "Produção"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Cookie de sessão gerado pelo NextAuth"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-INTERNAL-API-KEY",
        "description": "Chave de API para integrações externas (env X_INTERNAL_API_KEY)"
      }
    },
...
  },
  "paths": {
...
    "/api/payments/mercadopago/webhook": {
      "post": {
        "tags": ["Pagamentos"],
        "summary": "Webhook do Mercado Pago",
        "description": "Recebe notificações de pagamento. Valida assinatura HMAC e enfileira processamento via BullMQ.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string", "example": "payment" },
                  "data": { "type": "object", "properties": { "id": { "type": "string" } } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Recebido" } }
      }
    },
...
    "/api/coupons/validate": {
      "post": {
        "tags": ["Cupons"],
        "summary": "Validar e calcular desconto de um cupom",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code", "cartTotal"],
                "properties": {
                  "code": { "type": "string", "example": "DESCONTO10" },
                  "cartTotal": { "type": "number", "example": 250.00 },
                  "userId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado da validação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": { "type": "boolean" },
                    "couponId": { "type": "string" },
                    "discount": { "type": "number" },
                    "discountValue": { "type": "number" },
                    "discountType": { "type": "string", "enum": ["PERCENTAGE", "FIXED"] },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consent": {
      "post": {
        "tags": ["LGPD"],
        "summary": "Registrar consentimento de cookies (Art. 7 LGPD)",
        "description": "Chamado automaticamente pelo banner de cookies. Cria um registro auditável no banco. Não requer autenticação.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["analytics", "marketing", "version"],
                "properties": {
                  "essential": { "type": "boolean", "default": true },
                  "analytics": { "type": "boolean" },
                  "marketing": { "type": "boolean" },
                  "version": { "type": "string", "example": "1" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Consentimento registrado", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } } } } } }
      }
    },
...
    "/api/integrations/marketing/abandoned-carts": {
      "get": {
        "tags": ["Integrações"],
        "summary": "Listar carrinhos abandonados",
        "parameters": [{ "in": "query", "name": "hours", "schema": { "type": "integer", "default": 24 } }],
        "responses": { "200": { "description": "Carrinhos abandonados" } }
      },
      "post": {
        "tags": ["Integrações"],
        "summary": "Disparar e-mails de recuperação de carrinhos (CRON_SECRET)",
        "description": "Requer header `Authorization: Bearer <CRON_SECRET>`",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cartId": { "type": "string" },
                  "hours": { "type": "integer", "default": 24 }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "E-mails enviados", "content": { "application/json": { "schema": { "type": "object", "properties": { "sent": { "type": "integer" }, "failed": { "type": "integer" }, "total": { "type": "integer" } } } } } }
        }
      }
    },
...
  }
}